requirements.txt
appdirs==1.4.4
click==7.1.2
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
six==1.15.0
virtualenv==20.4.2
Werkzeug==1.0.1
flask_cors==3.0.10
Flask-JWT==0.3.2
bcrypt==3.2.0
pandas==1.2.1
dnspython==2.1.0
flask-sqlalchemy==2.4.4
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.24.2
pymysql==1.0.2